1. Added validation for title, author and year 2. Deleted extra controller 3. Implemented exception handling for constraints and not found cases 4. Created additional folder for BookServiceImpl class

+++ src/main/java/com/homework/crud/service/implementation/BookServiceImpl.java
@@ -1,8 +1,9 @@
-package com.homework.crud.service;
+package com.homework.crud.service.implementation;
 
+import com.homework.crud.exception.BookNotFoundException;
 import com.homework.crud.model.Book;
 import com.homework.crud.repository.BookRepository;
-import java.util.List;
+import com.homework.crud.service.BookService;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -10,18 +11,13 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 @Service
-public class BookServiceImpl implements BookService{
+public class BookServiceImpl implements BookService {
 
   private final BookRepository bookRepository;
 
   @Autowired
   public BookServiceImpl(BookRepository bookRepository) {
     this.bookRepository = bookRepository;
-  }
-
-  @Override
-  public List<Book> getAllBooks() {
-    return bookRepository.findAll();
   }
 
   @Override
@@ -36,7 +32,8 @@ public class BookServiceImpl implements BookService{
 
   @Override
   public Optional<Book> getBookById(int id) {
-    return bookRepository.findById(id);
+    return Optional.ofNullable(bookRepository.findById(id)
+        .orElseThrow(() -> new BookNotFoundException(id)));
   }
 
   @Override
@@ -46,11 +43,17 @@ public class BookServiceImpl implements BookService{
 
   @Override
   public Book updateBook(Book book) {
+    if (!bookRepository.existsById(book.getId())) {
+      throw new BookNotFoundException(book.getId());
+    }
     return bookRepository.save(book);
   }
 
   @Override
   public void deleteBook(int id) {
+    if (!bookRepository.existsById(id)) {
+      throw new BookNotFoundException(id);
+    }
     bookRepository.deleteById(id);
   }
 }
